Speech/requirements.txt
Flask
TTS
openai-whisper
soundfile